- Better recovery of bluetooth GPS after switching device off and on - Marked points appended to file 'xcsoar-marks.txt' - CDI display configurable - Settings->Display split into two pages - Sunset time shown in waypoint details - AAT and airspace areas drawn below waypoints and topology - Messagebox enhancements - MODIS Satelite images now co-located with waypoint file - Launcher now uninstalls/reinstalls properly. - Proper spelling of McCready (sorry, Paul!) - Display blanking automatically after one minute of UI inactivity if in   battery mode, reactivated with key press - New GPS status icons, less obtrusive. - Aircraft disappears when GPS is not connected - New "Status" summary page from main menu, giving aircraft position,   nearest waypoint range/bearing, local sunset time, GPS status - Additional airspace file can be specified for NOTAM airspace updates - Settings->File page split into two (map data separated off) - Snail trail toggles between no trail, long trail, and short trail

+++ Common/Source/Calculations.cpp
@@ -223,7 +223,7 @@ void Heading(NMEA_INFO *Basic, DERIVED_INFO *Calculated)
     }
 
     if (EnableCalibration) {
-      mag = isqrt4(x0*x0*100+y0*y0*100)/10.0;
+      mag = isqrt4((unsigned long)(x0*x0*100+y0*y0*100))/10.0;
       if ((Basic->AirspeedAvailable) && (Basic->IndicatedAirspeed>0)) {
 
         double k = (mag / Basic->TrueAirspeed);
